Remove hash_bundle_auth_data from OrchardPrimitives trait
